refactor: Separate views; improve clarity, maintainability; remove unused code
