+++ lib/models/reference_model.dart
@@ -7,7 +7,6 @@ class Reference implements Mappable {
   const Reference({required this.refId, required this.remark});
 
   /// Converts the Reference instance into a map.
-  /// This is required for the AppSelectionField to dynamically access its data.
   @override
   Map<String, dynamic> toMap() {
     return {
