+++ lib/screens/test_screens/fraud_menu.dart
@@ -1,8 +1,8 @@
 // --IMPORTANT : REMOVE THIS FILE LATER-- //
 
 import 'package:flutter/material.dart';
-import 'package:myapp/app_routes.dart'; 
-import 'package:myapp/services/icon_mapper.dart'; 
+import 'package:myapp/app_routes.dart';
+import 'package:myapp/services/icon_mapper.dart';
 import 'package:myapp/theme/app_theme.dart';
 import 'package:myapp/widgets/app_footer.dart'; // Your AppFooter
 
@@ -22,7 +22,7 @@ class FraudMenuScreen extends StatelessWidget {
       {
         'label': 'Attendance',
         'iconName': 'checklist',
-        'route': '/attendence', // CHANGE: Was AppRoutes.attendence
+        'route': '/attendance', // CHANGE: Was AppRoutes.attendance
       },
       {
         'label': 'Profile',
