+++ lib/views/region_selection_view.dart
@@ -3,15 +3,14 @@ import 'package:myapp/models/region_model.dart';
 import 'package:myapp/widgets/app_action_button.dart';
 import 'package:myapp/widgets/app_helper_field.dart';
 
+// Region Selection View
 class SelectRegionView extends StatefulWidget {
-  //final List<Region> Regions;
   final Function(Region) onRegionSelected;
   final VoidCallback onSubmit;
   final Region? selectedRegion;
 
   const SelectRegionView({
     super.key,
-    //required this.Regions,
     required this.onRegionSelected,
     required this.onSubmit,
     this.selectedRegion,
@@ -23,13 +22,11 @@ class SelectRegionView extends StatefulWidget {
 
 class _SelectRegionViewState extends State<SelectRegionView> {
   final TextEditingController _RegionController = TextEditingController();
-  // 1. This is the only state we need to track now.
   bool _isRegionSelectionCommitted = false;
 
   @override
   void initState() {
     super.initState();
-    // If a Region is pre-selected, the state is initially valid.
     if (widget.selectedRegion != null) {
       _RegionController.text = widget.selectedRegion!.region;
       _isRegionSelectionCommitted = true;
@@ -54,10 +51,7 @@ class _SelectRegionViewState extends State<SelectRegionView> {
             labelText: 'Select Region',
             selectionSheetTitle: 'Select a Region',
             initialValue: widget.selectedRegion,
-
-            //items: widget.Regions,
             onSelected: widget.onRegionSelected,
-            //displayString: (Region) => Region.name,
             onCommitStateChanged: (isCommitted) {
               setState(() {
                 _isRegionSelectionCommitted = isCommitted;
